Day 22 parts 1 and 2
I got some help from the Reddit thread for this one:

https://www.reddit.com/r/adventofcode/comments/5jor9q/2016_day_22_solutions/

For part 1, I got the right answer, but I was calculating it
the wrong way because I was checking the 'size' of node A instead of its
'used'. That meant that I thought that the idea of a viable pair wasn't
described correctly. I looked at reddit to see if anyone had mentioned
this, and saw several hints that helped with part b.